requirements.txt
tweepy
python-dotenv
google-cloud-storage